time_series_pipeline/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
statsmodels
prophet
tensorflow
reportlab
python-docx
openpyxl
streamlit
plotly
